feat: Add kt - Klingon Tools script for setting up git repos
The commit adds the `kt` script, which is a python script used to set up new git repositories and related tools in a uniform structured manner. It takes care of setting up directory and file structures, README files, and other common files using a question/answer format.

Note: This commit message follows the established convention of using a "feat" type for feature additions.
Signed-off-by: David Hooton <david@hooton.org>