chore: Update GitHub workflow to compile requirements if necessary and install dependencies
This commit modifies the GitHub workflow to include two new steps. First, it adds a step to compile requirements if the `requirements.in` file is not empty. This step uses the `pip-tools` package to generate a `requirements.txt` file based on the `requirements.in` file. Second, it adds a step to install dependencies if the `requirements.txt` file exists. This step upgrades `pip` and installs the dependencies listed in the `requirements.txt` file.

Signed-off-by: David Hooton <david@hooton.org>