chore: Compile requirements.txt in Makefile
The commit adds a new step to the Makefile to compile the requirements.txt file from the requirements.in file. This step uses the `pip-compile` command to generate the requirements.txt file based on the dependencies listed in requirements.in. This ensures that the project has an up-to-date and consistent list of dependencies.

Note: This commit message follows the established convention of using a "chore" type for miscellaneous changes.

Signed-off-by: David Hooton <david@hooton.org>