chore: Update GitHub workflow to use secrets for GITHUB_TOKEN
The commit updates the GitHub workflow to use the `secrets.GITHUB_TOKEN` instead of the `GITHUB_TOKEN` environment variable. This ensures that the token is securely stored and not exposed in the workflow file.

Note: This commit message follows the established convention of using a "chore" type for miscellaneous changes.

Signed-off-by: David Hooton <david@hooton.org>