chore: Add unit test workflow
This commit adds a new GitHub Actions workflow file, `unit-tests.yaml`, which sets up a job to run unit tests on the code. The workflow is triggered on pull requests to the `main` branch. It checks out the code, sets up Python, installs dependencies, and runs the tests using the `make` command.

Signed-off-by: David Hooton <david@hooton.org>