chore: Increment version number and update setup.py
This commit increments the version number and updates the setup.py file. It fetches the highest version number from PyPI, TestPyPI, the VERSION file, and setup.py, determines the highest version number, and increments it by 0.0.1. The new version number is echoed and saved in the VERSION file. The version number in setup.py is also updated to match the new version. Additionally, it checks if the VERSION and setup.py are in sync to avoid any discrepancies.

Signed-off-by: David Hooton <david@hooton.org>