chore: Update version number and push to origin
The `update-version` target in the Makefile has been modified to update the version number by incrementing the third (C) number by 1. It also validates the new version number and synchronizes it with the `setup.py` file. Additionally, it now pushes the updated version to the origin repository. This change ensures that the version number is accurately reflected and tracked in the repository.

Signed-off-by: David Hooton <david@hooton.org>