✨ feat(git_tools): allow custom commit message in git_commit_file
Updated the `git_commit_file` function in `git_tools.py` to accept a custom
commit message as an optional parameter. If a commit message is provided, the
function will commit the file with the specified message. If no message is
provided, a new commit message will be generated using OpenAI tools. Logs
successful commits and handles error scenarios accordingly.

Signed-off-by: David Hooton <david@hooton.org>