✨ feat(git_push):add ability to update submodule config file
Added functionality to check if the current repository is a submodule, stage
and commit changes in the submodule, navigate to the main repository, stage
and commit the updated submodule in the main repository, and push changes to
the main repository. If the current repository is not a submodule, the push
operation is performed at the end.

Signed-off-by: David Hooton <david@hooton.org>