⚙️ ci(workflows/release.yaml):update release workflow to create or edit pull requests based on branch comparison with GitHub Actions
Updated the release workflow in the `.github/workflows/release.yaml` file to
create or edit pull requests based on branch comparison using GitHub Actions.
The workflow now checks if a pull request already exists for the branch. If
not, it generates a new pull request title and creates a pull request with the
title and a standard body message. If a pull request already exists, it
fetches the existing title and updates the pull request body accordingly.

Signed-off-by: David Hooton <david@hooton.org>