📚 docs(workflow): add README for GitHub Actions workflows
This commit introduces a README.md file to the workflows 
directory, providing an overview and description of each GitHub 
Actions workflow used in the repository. The workflows covered 
include `auto-pr.yaml`, `run-tests.yaml`, `deploy.yaml`, and 
`release-drafter.yaml`, outlining their roles in the 
development and deployment process to PyPI. This documentation 
aims to enhance understanding and usage of the CI/CD processes.

Signed-off-by: David Hooton <david@hooton.org>